Move navigation list view tests to own file
